+++ src/main/java/com/jangburich/domain/owner/domain/controller/dto/res/OwnerGetResDTO.java
@@ -1,9 +1,11 @@
-package com.jangburich.domain.owner.domain;
+package com.jangburich.domain.owner.domain.controller.dto.res;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 import java.time.LocalDate;
+
+import com.jangburich.domain.owner.domain.entity.Owner;
 
 @Getter
 @RequiredArgsConstructor
