+++ src/main/java/com/jangburich/domain/owner/domain/repository/OwnerRepository.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.jangburich.domain.owner.domain.Owner;
+import com.jangburich.domain.owner.domain.entity.Owner;
 import com.jangburich.domain.user.domain.User;
 
 public interface OwnerRepository extends JpaRepository<Owner, Long> {
