+++ src/main/java/com/jangburich/domain/store/service/StoreService.java
@@ -4,7 +4,7 @@ import com.jangburich.domain.order.domain.OrderResponse;
 import com.jangburich.domain.order.domain.OrderStatus;
 import com.jangburich.domain.order.domain.Orders;
 import com.jangburich.domain.order.domain.repository.OrdersRepository;
-import com.jangburich.domain.owner.domain.Owner;
+import com.jangburich.domain.owner.domain.entity.Owner;
 import com.jangburich.domain.owner.domain.repository.OwnerRepository;
 import com.jangburich.domain.payment.domain.repository.TeamChargeHistoryRepository;
 import com.jangburich.domain.point.domain.TransactionType;
